feat: fix TS server SDK version
